// src/common/enums/order-status.enum.ts
export enum OrderStatus {
  PENDIENTE_PAGO = 'PENDIENTE_PAGO',
  CONFIRMADO = 'CONFIRMADO',
  EN_HORNEADA = 'EN_HORNEADA',
  EMPACADO = 'EMPACADO',
  EN_CAMINO = 'EN_CAMINO',
  LISTO_PARA_RETIRO = 'LISTO_PARA_RETIRO',
  ENTREGADO = 'ENTREGADO',
  CANCELADO = 'CANCELADO',
}

export enum DeliveryType {
  EXPRESS_LOCAL = 'EXPRESS_LOCAL',
  RETIRO_SUCURSAL = 'RETIRO_SUCURSAL',
  ENVIO_NACIONAL = 'ENVIO_NACIONAL',
}