// src/common/enums/payment-method.enum.ts
export enum PaymentMethod {
  QR_SIMPLE = 'QR_SIMPLE',
  TIGO_MONEY = 'TIGO_MONEY',
  TARJETA = 'TARJETA',
  EFECTIVO_CONTRAENTREGA = 'EFECTIVO_CONTRAENTREGA',
  TRANSFERENCIA_BANCARIA = 'TRANSFERENCIA_BANCARIA',
}

export enum PaymentStatus {
  PENDIENTE = 'PENDIENTE',
  PAGADO = 'PAGADO',
  RECHAZADO = 'RECHAZADO',
  REEMBOLSADO = 'REEMBOLSADO',
}